+++ src/angular/src/utils/params-list.ts
@@ -7,6 +7,7 @@ export const paramsList = [
   '_speed',
   'cssMode',
   'updateOnWindowResize',
+  'resizeObserver',
   'nested',
   '_width',
   '_height',
@@ -85,6 +86,9 @@ export const paramsList = [
   'slideDuplicatePrevClass',
   'wrapperClass',
   'runCallbacksOnInit',
+  'observer',
+  'observeParents',
+  'observeSlideChildren',
 
   // modules
   'a11y',