+++ src/components/core/classes/addClasses.js
@@ -2,8 +2,8 @@ function prepareClasses(entries, prefix) {
   const resultClasses = [];
   entries.forEach((item) => {
     if (typeof item === 'object') {
-      Object.entries(item).forEach(([classNames, condition]) => {
-        if (condition) {
+      Object.keys(item).forEach((classNames) => {
+        if (item[classNames]) {
           resultClasses.push(prefix + classNames);
         }
       });
@@ -31,7 +31,7 @@ export default function addClasses() {
     { 'ios': device.ios },
     { 'css-mode': params.cssMode },
   ], params.containerModifierClass);
-
-  $el.addClass([...classNames, ...suffixes].join(' '));
+  classNames.push(...suffixes);
+  $el.addClass([...classNames].join(' '));
   swiper.emitContainerClasses();
 }