+++ src/components/core/defaults.js
@@ -6,6 +6,7 @@ export default {
   speed: 300,
   cssMode: false,
   updateOnWindowResize: true,
+  resizeObserver: false,
   nested: false,
 
   // Overrides
@@ -37,6 +38,7 @@ export default {
 
   // Breakpoints
   breakpoints: undefined,
+  breakpointsBase: 'window',
 
   // Slides grid
   spaceBetween: 0,