+++ src/vue/update-swiper.js
@@ -30,7 +30,7 @@ function updateSwiper(swiper, slides, passedParams, changedParams) {
     changedParams.includes('pagination') &&
     passedParams.pagination &&
     passedParams.pagination.el &&
-    currentParams.pagination &&
+    (currentParams.pagination || currentParams.pagination === false) &&
     pagination &&
     !pagination.el
   ) {
@@ -41,7 +41,7 @@ function updateSwiper(swiper, slides, passedParams, changedParams) {
     changedParams.includes('scrollbar') &&
     passedParams.scrollbar &&
     passedParams.scrollbar.el &&
-    currentParams.scrollbar &&
+    (currentParams.scrollbar || currentParams.scrollbar === false) &&
     scrollbar &&
     !scrollbar.el
   ) {
@@ -53,7 +53,7 @@ function updateSwiper(swiper, slides, passedParams, changedParams) {
     passedParams.navigation &&
     passedParams.navigation.prevEl &&
     passedParams.navigation.nextEl &&
-    currentParams.navigation &&
+    (currentParams.navigation || currentParams.navigation === false) &&
     navigation &&
     !navigation.prevEl &&
     !navigation.nextEl
@@ -72,6 +72,8 @@ function updateSwiper(swiper, slides, passedParams, changedParams) {
   if (changedParams.includes('children') && virtual && currentParams.virtual.enabled) {
     virtual.slides = slides;
     virtual.update(true);
+  } else if (changedParams.includes('children') && swiper.lazy && swiper.params.lazy.enabled) {
+    swiper.lazy.load();
   }
 
   if (needThumbsInit) {